+++ python/onvo/onvo.py
@@ -1,6 +1,8 @@
 import os
 from onvo.accounts import Accounts
 
+# from onvo.automations import Automations
+
 try:
     default_endpoint = os.environ["ONVO_API_ENDPOINT"]
     default_api_key = os.environ["ONVO_API_KEY"]
@@ -18,10 +20,10 @@ class Onvo:
         params = [endpoint, api_key]
 
         self.accounts = Accounts(*params)
+        # self.automations = Automations.new(*params)
         # self.teams = Teams.new(*params)
         # self.embed_users = EmbedUsers.new(*params)
         # self.datasources = Datasources.new(*params)
-        # self.automations = Automations.new(*params)
         # self.dashboards = Dashboards.new(*params)
 
     def check_init_params(self):
